+++ src/CartServiceProvider.php
@@ -2,6 +2,10 @@
 
 namespace Masterix21\LaravelCart;
 
+use Livewire\Livewire;
+use Masterix21\LaravelCart\Http\Livewire\Counter;
+use Masterix21\LaravelCart\Http\Livewire\Items;
+use Masterix21\LaravelCart\View\IconDropdown;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
 
@@ -9,16 +13,32 @@ class CartServiceProvider extends PackageServiceProvider
 {
     public function configurePackage(Package $package): void
     {
-        /*
-         * This class is a Package Service Provider
-         *
-         * More info: https://github.com/spatie/laravel-package-tools
-         */
         $package
-            ->name('laravel-cart')
+            ->name('cart')
             ->hasConfigFile()
+            ->hasTranslations()
+            ->hasViews()
+            ->hasViewComponents('cart', IconDropdown::class)
             ->hasMigration('create_cart_items_table');
 
         $this->app->singleton('cart', fn () => new CartManager());
+    }
+
+    public function packageBooted(): void
+    {
+        $this->loadJsonTranslationsFrom($this->package->basePath('/../resources/lang/'));
+        $this->loadJsonTranslationsFrom(resource_path('lang/vendor/'. $this->package->name));
+
+        $this->registerLivewireComponents();
+    }
+
+    public function registerLivewireComponents(): self
+    {
+        if (class_exists(Livewire::class)) {
+            Livewire::component('cart-counter', Counter::class);
+            Livewire::component('cart-items', Items::class);
+        }
+
+        return $this;
     }
 }