+++ src/View/IconDropdown.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Masterix21\LaravelCart\View;
+
+use Illuminate\View\Component;
+
+class IconDropdown extends Component
+{
+    public function render()
+    {
+        return view('cart::icon-dropdown');
+    }
+}