+++ CHANGELOG.md
@@ -2,5 +2,18 @@
 
 All notable changes to `laravel-cart` will be documented in this file.
 
+## 1.0.2 - 2021-05-19 
+- 💅 Add `Cart::query`
+- 💅 Add `Cart::isEmpty`
+- 💅 Add `Cart::isNotEmpty`
+- 💅 Add `Cart::count`
+- 💅 Blade Component `<x-cart-icon-dropdown />`
+- 💅 Livewire component `<livewire:cart-counter />`
+- 💅 Livewire component `<livewire:cart-items />`
+- Italian translations
+
+## 1.0.1 - 2021-05-14
+- Composer fixes
+
 ## 1.0.0 - 2021-05-14
 - Initial release